Prefixe de class
