+++ Tests/Units/Planning/Creneau/CreneauController.php
@@ -1,42 +1,45 @@
 <?php
-namespace LibertAPI\Tests\Units\Planning;
+namespace LibertAPI\Tests\Units\Planning\Creneau;
 
-use \LibertAPI\Planning\Controller as _Controller;
+use LibertAPI\Planning\Creneau\CreneauController as _Controller;
 
 /**
- * Classe de test du contrôleur de planning
+ * Classe de test du contrôleur de créneau de planning
  *
  * @author Prytoegrian <prytoegrian@protonmail.com>
  * @author Wouldsmina
  *
  * @since 0.1
  */
-final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AController
+final class CreneauController extends \LibertAPI\Tests\Units\Tools\Libraries\AController
 {
     /**
-     * @var \LibertAPI\Planning\Repository Mock du repository associé
+     * @var \LibertAPI\Planning\Creneau\Repository Mock du repository associé
      */
     private $repository;
 
     /**
-     * @var \LibertAPI\Planning\Entite Mock de l'entité associée
+     * @var \LibertAPI\Planning\Creneau\Entite Mock de l'entité associée
      */
     private $entite;
 
     /**
      * Init des tests
      */
-    public function beforeTestMethod($method)
-    {
+    public function beforeTestMethod($method) {
         parent::beforeTestMethod($method);
         $this->mockGenerator->orphanize('__construct');
         $this->mockGenerator->shuntParentClassCalls();
-        $this->repository = new \mock\LibertAPI\Planning\Repository();
+        $this->repository = new \mock\LibertAPI\Planning\Creneau\CreneauRepository();
         $this->mockGenerator->orphanize('__construct');
-        $this->entite = new \mock\LibertAPI\Planning\Entite();
+        $this->entite = new \mock\LibertAPI\Planning\Creneau\CreneauEntite();
         $this->entite->getMockController()->getId = 42;
-        $this->entite->getMockController()->getName = 12;
-        $this->entite->getMockController()->getStatus = 12;
+        $this->entite->getMockController()->getPlanningId = 12;
+        $this->entite->getMockController()->getJourId = 12;
+        $this->entite->getMockController()->getTypeSemaine = 12;
+        $this->entite->getMockController()->getTypePeriode = 12;
+        $this->entite->getMockController()->getDebut = 12;
+        $this->entite->getMockController()->getFin = 12;
     }
 
     /*************************************************
@@ -51,7 +54,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         $this->repository->getMockController()->getOne = $this->entite;
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->get($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->get($this->request, $this->response, ['creneauId' => 99, 'planningId' => 45]);
         $data = $this->getJsonDecoded($response->getBody());
 
         $this->integer($response->getStatusCode())->isIdenticalTo(200);
@@ -73,7 +76,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         };
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->get($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->get($this->request, $this->response, ['creneauId' => 99, 'planningId' => 45]);
 
         $this->assertError($response, 404);
     }
@@ -89,7 +92,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         $controller = new _Controller($this->repository, $this->router);
 
         $this->exception(function () use ($controller) {
-            $controller->get($this->request, $this->response, ['planningId' => 99]);
+            $controller->get($this->request, $this->response, ['creneauId' => 99, 'planningId' => 45]);
         })->isInstanceOf('\Exception');
     }
 
@@ -98,13 +101,12 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
      */
     public function testGetListFound()
     {
-        $this->request->getMockController()->getQueryParams = [];
         $this->repository->getMockController()->getList = [
             42 => $this->entite,
         ];
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->get($this->request, $this->response, []);
+        $response = $controller->get($this->request, $this->response, ['planningId' => 45]);
         $data = $this->getJsonDecoded($response->getBody());
 
         $this->integer($response->getStatusCode())->isIdenticalTo(200);
@@ -112,7 +114,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
             ->integer['code']->isIdenticalTo(200)
             ->string['status']->isIdenticalTo('success')
             ->string['message']->isIdenticalTo('')
-            //->array['data']->hasSize(1) // TODO: l'asserter atoum en sucre syntaxique est buggé, faire un ticket
+            //->array['data']->hasSize(1) // l'asserter atoum en sucre syntaxique est buggé, faire un ticket
         ;
         $this->array($data['data'][0])->hasKey('id');
     }
@@ -122,13 +124,12 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
      */
     public function testGetListNotFound()
     {
-        $this->request->getMockController()->getQueryParams = [];
         $this->repository->getMockController()->getList = function () {
             throw new \UnexpectedValueException('');
         };
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->get($this->request, $this->response, []);
+        $response = $controller->get($this->request, $this->response, ['planningId' => 45]);
 
         $this->assertSuccessEmpty($response);
     }
@@ -138,14 +139,13 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
      */
     public function testGetListFallback()
     {
-        $this->request->getMockController()->getQueryParams = [];
         $this->repository->getMockController()->getList = function () {
             throw new \Exception('');
         };
         $controller = new _Controller($this->repository, $this->router);
 
         $this->exception(function () use ($controller) {
-            $controller->get($this->request, $this->response, []);
+            $controller->get($this->request, $this->response, ['planningId' => 45]);
         })->isInstanceOf('\Exception');
     }
 
@@ -153,10 +153,12 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
      * POST
      *************************************************/
 
+    // post ok
+
     /**
      * Teste la méthode post d'un json mal formé
      */
-    public function testPostJsonBadFormat()
+    public function testPostJsonBadformat()
     {
         // Le framework fait du traitement, un mauvais json est simplement null
         $this->request->getMockController()->getParsedBody = null;
@@ -170,10 +172,10 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
     /**
      * Teste la méthode post avec un argument de body manquant
      */
-    public function testPostMissingRequiredArg()
+    public function testPostMissingArgument()
     {
-        $this->request->getMockController()->getParsedBody = [];
-        $this->repository->getMockController()->postOne = function () {
+        $this->request->getMockController()->getParsedBody = [[]];
+        $this->repository->getMockController()->postList = function () {
             throw new \LibertAPI\Tools\Exceptions\MissingArgumentException('');
         };
         $controller = new _Controller($this->repository, $this->router);
@@ -188,9 +190,9 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
      */
     public function testPostBadDomain()
     {
-        $this->request->getMockController()->getParsedBody = [];
-        $this->repository->getMockController()->postOne = function () {
-            throw new \DomainException('Status doit être un int');
+        $this->request->getMockController()->getParsedBody = [[]];
+        $this->repository->getMockController()->postList = function () {
+            throw new \DomainException('');
         };
         $controller = new _Controller($this->repository, $this->router);
 
@@ -200,13 +202,28 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
     }
 
     /**
-     * Teste la méthode post Ok
+     * Teste le fallback de la méthode post
+     */
+    public function testPostFallback()
+    {
+        $this->request->getMockController()->getParsedBody = [[]];
+        $this->repository->getMockController()->postList = function () {
+            throw new \LogicException('');
+        };
+        $controller = new _Controller($this->repository, $this->router);
+
+        $this->exception(function () use ($controller) {
+            $controller->post($this->request, $this->response);
+        })->isInstanceOf('\Exception');
+    }
+    /**
+     * Teste la méthode post tout ok
      */
     public function testPostOk()
     {
-        $this->request->getMockController()->getParsedBody = [];
+        $this->request->getMockController()->getParsedBody = [[]];
         $this->router->getMockController()->pathFor = '';
-        $this->repository->getMockController()->postOne = 42;
+        $this->repository->getMockController()->postList = [42, 74, 314];
         $controller = new _Controller($this->repository, $this->router);
 
         $response = $controller->post($this->request, $this->response);
@@ -219,22 +236,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
             ->string['message']->isIdenticalTo('')
             ->array['data']->isNotEmpty()
         ;
-    }
-
-    /**
-     * Teste le fallback de la méthode post
-     */
-    public function testPostFallback()
-    {
-        $this->request->getMockController()->getParsedBody = [];
-        $this->repository->getMockController()->postOne = function () {
-            throw new \Exception('');
-        };
-        $controller = new _Controller($this->repository, $this->router);
-
-        $this->exception(function () use ($controller) {
-            $controller->post($this->request, $this->response);
-        })->isInstanceOf('\Exception');
+        $this->integer(count($data['data']))->isIdenticalTo(3);
     }
 
     /*************************************************
@@ -266,7 +268,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         };
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->put($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->put($this->request, $this->response, ['creneauId' => 99]);
 
         $this->boolean($response->isNotFound())->isTrue();
     }
@@ -283,7 +285,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         $controller = new _Controller($this->repository, $this->router);
 
         $this->exception(function () use ($controller) {
-            $controller->put($this->request, $this->response, ['planningId' => 99]);
+            $controller->put($this->request, $this->response, ['creneauId' => 99]);
         })->isInstanceOf('\Exception');
     }
 
@@ -300,7 +302,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         };
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->put($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->put($this->request, $this->response, ['creneauId' => 99]);
 
         $this->assertError($response, 412);
     }
@@ -317,7 +319,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         };
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->put($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->put($this->request, $this->response, ['creneauId' => 99]);
 
         $this->assertError($response, 412);
     }
@@ -335,7 +337,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         $controller = new _Controller($this->repository, $this->router);
 
         $this->exception(function () use ($controller) {
-            $controller->put($this->request, $this->response, ['planningId' => 99]);
+            $controller->put($this->request, $this->response, ['creneauId' => 99]);
         })->isInstanceOf('\Exception');
     }
 
@@ -349,7 +351,7 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
         $this->repository->getMockController()->putOne = '';
         $controller = new _Controller($this->repository, $this->router);
 
-        $response = $controller->put($this->request, $this->response, ['planningId' => 99]);
+        $response = $controller->put($this->request, $this->response, ['creneauId' => 99]);
 
         $data = $this->getJsonDecoded($response->getBody());
 
@@ -359,60 +361,6 @@ final class Controller extends \LibertAPI\Tests\Units\Tools\Libraries\AControlle
             ->string['status']->isIdenticalTo('success')
             ->string['message']->isIdenticalTo('')
             ->string['data']->isIdenticalTo('')
-        ;
-    }
-
-    /*************************************************
-     * DELETE
-     *************************************************/
-
-    /**
-     * Teste la méthode delete avec un détail non trouvé (id en Bad domaine)
-     */
-    public function testDeleteNotFound()
-    {
-        $this->repository->getMockController()->getOne = function () {
-            throw new \DomainException('');
-        };
-        $controller = new _Controller($this->repository, $this->router);
-
-        $response = $controller->delete($this->request, $this->response, ['planningId' => 99]);
-
-        $this->boolean($response->isNotFound())->isTrue();
-    }
-
-    /**
-     * Teste le fallback de la méthode delete
-     */
-    public function testDeleteFallback()
-    {
-        $this->repository->getMockController()->getOne = function () {
-            throw new \LogicException('');
-        };
-        $controller = new _Controller($this->repository, $this->router);
-
-        $this->exception(function () use ($controller) {
-            $controller->delete($this->request, $this->response, ['planningId' => 99]);
-        })->isInstanceOf('\Exception');
-    }
-
-    /**
-     * Teste la méthode delete Ok
-     */
-    public function testDeleteOk()
-    {
-        $this->repository->getMockController()->getOne = $this->entite;
-        $controller = new _Controller($this->repository, $this->router);
-
-        $response = $controller->delete($this->request, $this->response, ['planningId' => 99]);
-        $data = $this->getJsonDecoded($response->getBody());
-
-        $this->integer($response->getStatusCode())->isIdenticalTo(200);
-        $this->array($data)
-            ->integer['code']->isIdenticalTo(200)
-            ->string['status']->isIdenticalTo('success')
-            ->string['message']->isIdenticalTo('')
-            ->array['data']->isNotEmpty()
         ;
     }
 }
